+++ src/generator/settings.ts
@@ -4,14 +4,11 @@ import * as vscode from 'vscode';
 export const getSetting_customTemplateFolder = () =>
   getExtensionSetting<string | null | undefined>('customTemplateFolder');
 
-export const getSetting_defaultSpecsUseTestBed = () =>
-  getExtensionSetting<boolean | undefined>('defaultSpecsUseTestBed');
-
 export const getSetting_generateSpec = () =>
-  getExtensionSetting<boolean | undefined>('generateSpec');
+  getExtensionSetting<number | undefined>('generateSpec');
 
 export const getSetting_generateStories = () =>
-  getExtensionSetting<boolean | undefined>('generateStories');
+  getExtensionSetting<number | undefined>('generateStories');
 
 function getExtensionSetting<T>(settingKey: string): T | undefined {
   return getExtensionSettings().get(settingKey);